+++ src/api/registration/ProjectOperationRegistration.ts
@@ -14,13 +14,29 @@
  * limitations under the License.
  */
 
-import { AnyProjectEditor } from "@atomist/automation-client/operations/edit/projectEditor";
+import { AnyProjectEditor, SimpleProjectEditor } from "@atomist/automation-client/operations/edit/projectEditor";
+import { chainTransforms } from "./CodeTransformRegistration";
 import { CommandRegistration } from "./CommandRegistration";
 
 /**
  * Function that can transform a project
  */
-export type CodeTransform<P = any> = AnyProjectEditor<P>;
+export type CodeTransform<P = any> = SimpleProjectEditor<P>;
+
+export type CodeTransformRegisterable<P = any> = AnyProjectEditor<P>;
+
+/**
+ * One or many CodeTransforms
+ */
+export type CodeTransformOrTransforms<PARAMS> = CodeTransformRegisterable<PARAMS> | Array<CodeTransformRegisterable<PARAMS>>;
+
+export function toCodeTransformRegisterable<PARAMS>(ctot: CodeTransformOrTransforms<PARAMS>): CodeTransformRegisterable<PARAMS> {
+    if (Array.isArray(ctot)) {
+        return chainTransforms(...ctot);
+    } else {
+        return ctot as CodeTransform<PARAMS>;
+    }
+}
 
 /**
  * Superclass for all registrations of "project operations",
@@ -32,7 +48,7 @@ export interface ProjectOperationRegistration<PARAMS> extends CommandRegistratio
     /**
      * Function to transform the project
      */
-    transform?: CodeTransform<PARAMS> | Array<CodeTransform<PARAMS>>;
+    transform?: CodeTransformOrTransforms<PARAMS>;
 
     /**
      * Create the editor function that can modify a project